requirements.txt
numpy>=2.3.0
pandas>=2.3.0
SQLAlchemy>=2.0.40
transformers>=4.44.0
huggingface-hub>=0.24.0
python-dotenv>=1.0.0
pytest>=8.2.0
rich>=13.7.0
